requirement.txt
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.3
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.2.1
Django==2.0.4
dlib==19.6.1
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lxml==4.2.1
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numpy==1.14.2
opencv-python==3.4.0.12
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyWavelets==0.5.2
pywinpty==0.5.1
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
scikit-image==0.13.1
scipy==1.0.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
